feat: schema codegen bucketing to help with tree-shaking
