+++ private/smithy-rpcv2-cbor-schema/src/commands/FractionalSecondsCommand.ts
@@ -2,7 +2,7 @@
 import { RpcV2ProtocolClientResolvedConfig, ServiceInputTypes, ServiceOutputTypes } from "../RpcV2ProtocolClient";
 import { commonParams } from "../endpoint/EndpointParameters";
 import { FractionalSecondsOutput } from "../models/models_0";
-import { FractionalSeconds } from "../schemas/schemas_2_FractionalSeconds";
+import { FractionalSeconds } from "../schemas/schemas_3_FractionalSeconds";
 import { getEndpointPlugin } from "@smithy/middleware-endpoint";
 import { Command as $Command } from "@smithy/smithy-client";
 import { MetadataBearer as __MetadataBearer } from "@smithy/types";
