+++ private/smithy-rpcv2-cbor-schema/src/commands/NoInputOutputCommand.ts
@@ -1,7 +1,7 @@
 // smithy-typescript generated code
 import { RpcV2ProtocolClientResolvedConfig, ServiceInputTypes, ServiceOutputTypes } from "../RpcV2ProtocolClient";
 import { commonParams } from "../endpoint/EndpointParameters";
-import { NoInputOutput } from "../schemas/schemas_4_NoInputOutput";
+import { NoInputOutput } from "../schemas/schemas_5_NoInputOutput";
 import { getEndpointPlugin } from "@smithy/middleware-endpoint";
 import { Command as $Command } from "@smithy/smithy-client";
 import { MetadataBearer as __MetadataBearer } from "@smithy/types";
