+++ private/smithy-rpcv2-cbor-schema/src/commands/SparseNullsOperationCommand.ts
@@ -2,7 +2,7 @@
 import { RpcV2ProtocolClientResolvedConfig, ServiceInputTypes, ServiceOutputTypes } from "../RpcV2ProtocolClient";
 import { commonParams } from "../endpoint/EndpointParameters";
 import { SparseNullsOperationInputOutput } from "../models/models_0";
-import { SparseNullsOperation } from "../schemas/schemas_0";
+import { SparseNullsOperation } from "../schemas/schemas_1_Rpc";
 import { getEndpointPlugin } from "@smithy/middleware-endpoint";
 import { Command as $Command } from "@smithy/smithy-client";
 import { MetadataBearer as __MetadataBearer } from "@smithy/types";
