+++ private/smithy-rpcv2-cbor-schema/src/schemas/schemas_2_EmptyInputOutput.ts
