+++ private/smithy-rpcv2-cbor-schema/src/schemas/schemas_4_.ts
@@ -1,5 +1,6 @@
 // smithy-typescript generated code
-import { Unit, _FO, _Fl, _v, n1 } from "./schemas_0";
+import { _FO, _Fl, _v, n1 } from "./schemas_0";
+import { Unit } from "./schemas_1_Rpc";
 import { op, struct } from "@smithy/core/schema";
 
 /* eslint no-var: 0 */
